api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.34
pymysql==1.1.1
mysqlclient==2.2.4 
python-dotenv==1.0.1
pydantic-settings==2.4.0
alembic==1.13.2   
requests==2.32.3 